+++ tests/test_hello.py
@@ -19,3 +19,31 @@ def test_single_container(docker_hello_world, docker_services):
     res = docker_services._docker_compose.execute("ps")
     assert 'hello' in res
     assert 'hello2' not in res
+
+
+def test_exec(docker_services):
+    """Test the exec method.
+
+    The exec method executes a command inside a docker command.
+    """
+    res = docker_services.exec('hello', 'ls', '-a')
+    assert res == '.\n..\nindex.html\n'
+
+
+# counter
+custom_checker_called = 0
+
+
+def custom_checker(ip_address, port):
+    global custom_checker_called
+    if custom_checker_called > 1:
+        return True
+    custom_checker_called += 1
+    return False
+
+
+def test_custom_checker(docker_services):
+    """Test a custom checker in the wait_for_service method."""
+
+    docker_services.wait_for_service("hello", 80, check_server=custom_checker)
+    assert custom_checker_called > 1