+++ CHANGES.rst
@@ -5,6 +5,9 @@ Changes for Lovely Pytest Docker
 unreleased
 ==========
 
+ - added exec method in Services class
+ - added possibility to use custom checker for wait_for_service method
+
 2018/01/09 0.0.1
 ================